Trim wasm-pkg-common to allow targetting wasm32-wasip1
