+++ crates/wasm-pkg-common/Cargo.toml
@@ -9,32 +9,27 @@ license.workspace = true
 readme = "../../README.md"
 
 [features]
-metadata-client = ["dep:reqwest"]
-tokio = ["tokio/io-util"]
+registry-config = [
+    "dep:etcetera",
+    "dep:tokio",
+    "dep:toml",
+]
 # Extra features to facilitate making working with OCI images easier
 oci_extras = []
 
 [dependencies]
 anyhow = { workspace = true }
 bytes = { workspace = true }
-etcetera = { workspace = true }
+etcetera = { workspace = true, optional = true }
 futures-util = { workspace = true }
 http = "1.1.0"
-reqwest = { version = "0.12.0", default-features = false, features = [
-    "rustls-tls",
-    "charset",
-    "http2",
-    "macos-system-configuration",
-    "json",
-], optional = true }
 semver = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 sha2 = { workspace = true }
-tokio = { workspace = true, features = ["fs"] }
-toml = { workspace = true }
+tokio = { workspace = true, optional = true, features = ["fs"] }
+toml = { workspace = true, optional = true }
 thiserror = { workspace = true }
-tracing = { workspace = true }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["macros", "rt"] }
