+++ crates/wasm-pkg-common/src/digest.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "tokio")]
-use std::path::Path;
 use std::str::FromStr;
 
 use bytes::Bytes;
@@ -16,22 +14,6 @@ pub enum ContentDigest {
 }
 
 impl ContentDigest {
-    #[cfg(feature = "tokio")]
-    pub async fn sha256_from_file(path: impl AsRef<Path>) -> Result<Self, std::io::Error> {
-        use tokio::io::AsyncReadExt;
-        let mut file = tokio::fs::File::open(path).await?;
-        let mut hasher = Sha256::new();
-        let mut buf = [0; 4096];
-        loop {
-            let n = file.read(&mut buf).await?;
-            if n == 0 {
-                break;
-            }
-            hasher.update(&buf[..n]);
-        }
-        Ok(hasher.into())
-    }
-
     pub fn validating_stream(
         &self,
         stream: impl TryStream<Ok = Bytes, Error = Error>,
