+++ crates/wkg/tests/e2e.rs
@@ -1,9 +1,6 @@
 mod common;
 
-#[cfg(any(target_os = "linux", feature = "_local"))]
-// NOTE: These are only run on linux for CI purposes, because they rely on the docker client being
-// available, and for various reasons this has proven to be problematic on both the Windows and
-// MacOS runners due to it not being installed (yay licensing).
+#[cfg(feature = "docker-tests")]
 #[tokio::test]
 async fn build_and_publish_with_metadata() {
     use oci_client::{client::ClientConfig, manifest::OciManifest, Reference};
